backend/requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
dj-database-url==2.2.0
psycopg2-binary==2.9.10
gunicorn==21.2.0
clerk-sdk==0.5.3
